vana class 4: delineare signed radii and dead clamps
Circle/ellipse radius params and praecisionem x/y go s32 (the dead
negative checks come alive; wrapped radius was a 4e9-iteration
hang). Circle midpoint error term was unsigned, degrading its <= 0
branch to == 0 - circles were subtly misshapen. Palette gray
distance was unsigned, so brighter palette entries could never win
nearest-gray. Rounded-rect dead lower clamps dropped (upper clamps
already bound wrapped values).

+++ lib/elementa.c
@@ -202,12 +202,12 @@ elementa_slider(
     delineare_lineam_horizontalem(ctx, px, px + track_latitudo, track_y, color_track);
 
     /* Delineare thumb */
-    delineare_circulum_plenum(ctx, thumb_x, track_y, thumb_radius, color_thumb);
+    delineare_circulum_plenum(ctx, thumb_x, track_y, (s32)thumb_radius, color_thumb);
 
     /* Si hover vel dragging, delineare circulum extra */
     si (fructus.hover || fructus.dragging)
     {
-        delineare_circulum(ctx, thumb_x, track_y, thumb_radius + 2, color_thumb);
+        delineare_circulum(ctx, thumb_x, track_y, (s32)(thumb_radius + 2), color_thumb);
     }
 
     redde fructus;
